refactor: migrate @suite-native/staking to @suite-common/wallet-core/staking and @suite-native/module-earn
