+++ suite-common/wallet-core/src/index.ts
@@ -75,6 +75,7 @@ export * from './staking/ethereum/ethereumNetworkAddresses';
 export * from './staking/ethereum/ethereumStaking';
 export * from './staking/ethereum/ethereumStakingConstants';
 export * from './staking/ethereum/ethereumStakingSelectors';
+export * from './staking/ethereum/ethereumStakingSignUtils';
 export * from './staking/ethereum/ethereumStakingUtils';
 export type * from './staking/ethereum/types';
 export * from './staking/hooks/useStakingEntryPeriodEstimateInDays';
@@ -84,6 +85,7 @@ export * from './staking/shared/stakingUtils';
 export * from './staking/solana/solanaStakingFormUtils';
 export * from './staking/solana/solanaStakingLimitUtils';
 export * from './staking/solana/solanaStakingSelectors';
+export * from './staking/solana/solanaStakingThunks';
 export * from './staking/solana/solanaStakingUtils';
 export * from './staking/stakingActions';
 export * from './staking/stakingConstants';
@@ -94,6 +96,7 @@ export type * from './staking/stakingReducerTypes';
 export * from './staking/stakingSelectors';
 export * from './staking/shared/stakingSelectors';
 export * from './staking/stakingThunks';
+export * from './staking/stakingTransactionThunks';
 export { composeTronClaimFeeLevelsThunk } from './staking/tron/actions/claim/composeClaim';
 export { submitTronClaimThunk } from './staking/tron/actions/claim/submitClaim';
 export { composeTronFreezeFeeLevelsThunk } from './staking/tron/actions/freeze/composeFreeze';
