+++ suite-native/module-earn/src/thunks/staking/solanaStakingThunks.test.ts
@@ -12,10 +12,7 @@ import {
 } from '@suite-common/wallet-types';
 import TrezorConnect from '@trezor/connect';
 
-import {
-    composeSolanaStakingTransactionFeeLevelsNativeThunk,
-    signSolanaStakingTransactionNativeThunk,
-} from './stakeFormSolanaNativeThunks';
+import { signSolanaStakingTransactionThunk } from './solanaStakingThunks';
 
 jest.mock('@trezor/connect', () => ({
     __esModule: true,
@@ -74,7 +71,6 @@ jest.mock('@suite-native/device-mutex', () => ({
 
 const STATIC_SESSION_ID = '1stTestnetAddress@device_id:0';
 const SOL_ACCOUNT_KEY = 'sol1' as AccountKey;
-const DSOL_ACCOUNT_KEY = 'dsol1' as AccountKey;
 
 const solAccount: Account = {
     symbol: 'sol',
@@ -85,12 +81,6 @@ const solAccount: Account = {
     path: "m/44'/501'/0'/0'",
     availableBalance: '10000000000',
     visible: true,
-} as unknown as Account;
-
-const dsolAccount: Account = {
-    ...solAccount,
-    symbol: 'dsol',
-    key: DSOL_ACCOUNT_KEY,
 } as unknown as Account;
 
 const solanaFeeBucket = {
@@ -155,25 +145,12 @@ const buildSolanaPrecomposedTransaction = (): PrecomposedTransactionFinal =>
 
 const solanaSignTransactionMock = TrezorConnect.solanaSignTransaction as jest.Mock;
 const blockchainGetInfoMock = TrezorConnect.blockchainGetInfo as jest.Mock;
-const blockchainEstimateFeeMock = TrezorConnect.blockchainEstimateFee as jest.Mock;
-
-const dispatchCompose = async (
-    store: ReturnType<typeof buildStore>,
-    args: Parameters<typeof composeSolanaStakingTransactionFeeLevelsNativeThunk>[0],
-) => {
-    const action = await store.dispatch(
-        composeSolanaStakingTransactionFeeLevelsNativeThunk(args) as any,
-    );
-    if (isFulfilled(action)) return { ok: true as const, payload: action.payload };
-    if (isRejected(action)) return { ok: false as const, error: action.payload };
-    throw new Error('Unexpected dispatch outcome');
-};
 
 const dispatchSign = async (
     store: ReturnType<typeof buildStore>,
-    args: Parameters<typeof signSolanaStakingTransactionNativeThunk>[0],
+    args: Parameters<typeof signSolanaStakingTransactionThunk>[0],
 ) => {
-    const action = await store.dispatch(signSolanaStakingTransactionNativeThunk(args) as any);
+    const action = await store.dispatch(signSolanaStakingTransactionThunk(args) as any);
     if (isFulfilled(action)) return { ok: true as const };
     if (isRejected(action)) return { ok: false as const, error: action.payload };
     throw new Error('Unexpected dispatch outcome');
@@ -192,8 +169,6 @@ beforeEach(() => {
         success: false,
         payload: { error: 'backend not connected' },
     });
-
-    blockchainEstimateFeeMock.mockClear();
 
     prepareStakeSolTxMock.mockReset();
     prepareStakeSolTxMock.mockResolvedValue({
@@ -215,179 +190,7 @@ beforeEach(() => {
     });
 });
 
-describe('composeSolanaStakingTransactionFeeLevelsNativeThunk', () => {
-    it('returns undefined for an empty amount', async () => {
-        const store = buildStore();
-
-        const result = await dispatchCompose(store, {
-            accountKey: SOL_ACCOUNT_KEY,
-            stakeType: 'stake',
-            amount: '',
-        });
-
-        expect(result).toEqual({ ok: true, payload: undefined });
-    });
-
-    it('composes a dsol (devnet) solana account the same way as sol', async () => {
-        const store = buildStore({ accounts: [dsolAccount] });
-
-        const result = await dispatchCompose(store, {
-            accountKey: DSOL_ACCOUNT_KEY,
-            stakeType: 'stake',
-            amount: '1',
-        });
-
-        expect(result.ok).toBe(true);
-        const levels = (result as { payload: Record<string, any> }).payload;
-        expect(levels.normal.type).toBe('final');
-        expect(levels.normal.solanaTxMeta.feeIncludingRentLamports).toBe('2287880');
-    });
-
-    it('composes an unstake the same way as stake', async () => {
-        const store = buildStore();
-
-        const result = await dispatchCompose(store, {
-            accountKey: SOL_ACCOUNT_KEY,
-            stakeType: 'unstake',
-            amount: '1',
-        });
-
-        expect(result.ok).toBe(true);
-        const levels = (result as { payload: Record<string, any> }).payload;
-        expect(levels.normal.type).toBe('final');
-        expect(levels.normal.solanaTxMeta.feeIncludingRentLamports).toBe('2287880');
-    });
-
-    it('composes a claim through prepareClaimSolTx (not the stake builder)', async () => {
-        const store = buildStore();
-
-        const result = await dispatchCompose(store, {
-            accountKey: SOL_ACCOUNT_KEY,
-            stakeType: 'claim',
-            amount: '1',
-        });
-
-        expect(result.ok).toBe(true);
-        const levels = (result as { payload: Record<string, any> }).payload;
-        expect(levels.normal.type).toBe('final');
-        expect(prepareClaimSolTxMock).toHaveBeenCalled();
-        expect(prepareStakeSolTxMock).not.toHaveBeenCalled();
-    });
-
-    it('returns precomposed fee levels with the solana tx meta applied', async () => {
-        const store = buildStore();
-
-        const result = await dispatchCompose(store, {
-            accountKey: SOL_ACCOUNT_KEY,
-            stakeType: 'stake',
-            amount: '1',
-        });
-
-        expect(result.ok).toBe(true);
-        const levels = (result as { payload: Record<string, any> }).payload;
-        expect(levels.normal.type).toBe('final');
-        // applySolanaTxMeta overrides the fee with feeIncludingRentLamports
-        expect(levels.normal.fee).toBe('2287880');
-        expect(levels.normal.solanaTxMeta.feeIncludingRentLamports).toBe('2287880');
-    });
-
-    it('omits newAccountProgramName when an unstake only deactivates (creates no stake account)', async () => {
-        // A plain deactivate reserves no rent, so the fee estimate must not add a phantom rent reserve.
-        prepareUnstakeSolTxMock.mockResolvedValue({
-            success: true,
-            txShim: solanaTxShim,
-            solanaTxMeta: { ...solanaTxMetaMock, rentLamports: '0' },
-        });
-        const store = buildStore();
-
-        await dispatchCompose(store, {
-            accountKey: SOL_ACCOUNT_KEY,
-            stakeType: 'unstake',
-            amount: '1',
-        });
-
-        expect(blockchainEstimateFeeMock).toHaveBeenCalled();
-        const request = blockchainEstimateFeeMock.mock.calls.at(-1)?.[0];
-        expect(request.request.specific).not.toHaveProperty('newAccountProgramName');
-    });
-
-    it('marks a transaction with a split instruction as device review only', async () => {
-        // A split instruction cannot be recreated in Suite, the whole review happens on the device.
-        prepareUnstakeSolTxMock.mockResolvedValue({
-            success: true,
-            txShim: solanaTxShim,
-            solanaTxMeta: { ...solanaTxMetaMock, hasSplitInstruction: true },
-        });
-        const store = buildStore();
-
-        const result = await dispatchCompose(store, {
-            accountKey: SOL_ACCOUNT_KEY,
-            stakeType: 'unstake',
-            amount: '1',
-        });
-
-        expect(result.ok).toBe(true);
-        const levels = (result as { payload: Record<string, any> }).payload;
-        expect(levels.normal.isDeviceReviewOnly).toBe(true);
-    });
-
-    it('keeps the device review only flag when the fee-aware preparation fails', async () => {
-        // The fee-aware rebuild is only a refinement, losing it must not turn a split into a Suite review.
-        prepareUnstakeSolTxMock.mockReset();
-        prepareUnstakeSolTxMock.mockResolvedValueOnce({
-            success: true,
-            txShim: solanaTxShim,
-            solanaTxMeta: { ...solanaTxMetaMock, hasSplitInstruction: true },
-        });
-        prepareUnstakeSolTxMock.mockResolvedValueOnce({ success: false });
-        const store = buildStore();
-
-        const result = await dispatchCompose(store, {
-            accountKey: SOL_ACCOUNT_KEY,
-            stakeType: 'unstake',
-            amount: '1',
-        });
-
-        expect(result.ok).toBe(true);
-        const levels = (result as { payload: Record<string, any> }).payload;
-        expect(levels.normal.isDeviceReviewOnly).toBe(true);
-    });
-
-    it('leaves a transaction without a split instruction reviewable in Suite', async () => {
-        const store = buildStore();
-
-        const result = await dispatchCompose(store, {
-            accountKey: SOL_ACCOUNT_KEY,
-            stakeType: 'unstake',
-            amount: '1',
-        });
-
-        expect(result.ok).toBe(true);
-        const levels = (result as { payload: Record<string, any> }).payload;
-        expect(levels.normal.isDeviceReviewOnly).toBe(false);
-    });
-
-    it('passes newAccountProgramName when an unstake splits a stake account (reserves rent)', async () => {
-        // A split creates a new rent-exempt stake account, so that reserve must be counted.
-        prepareUnstakeSolTxMock.mockResolvedValue({
-            success: true,
-            txShim: solanaTxShim,
-            solanaTxMeta: { ...solanaTxMetaMock, rentLamports: '2282880' },
-        });
-        const store = buildStore();
-
-        await dispatchCompose(store, {
-            accountKey: SOL_ACCOUNT_KEY,
-            stakeType: 'unstake',
-            amount: '1',
-        });
-
-        const request = blockchainEstimateFeeMock.mock.calls.at(-1)?.[0];
-        expect(request.request.specific.newAccountProgramName).toBe('staking');
-    });
-});
-
-describe('signSolanaStakingTransactionNativeThunk', () => {
+describe('signSolanaStakingTransactionThunk', () => {
     it('rejects when the account is not found', async () => {
         const errorSpy = jest.spyOn(console, 'error').mockImplementation(() => {});
         const store = buildStore({ accounts: [] });
