--- suite-native/staking/jest.config.js
@@ -1,5 +1,0 @@
-const baseConfig = require('../../jest.config.native');
-
-module.exports = {
-    ...baseConfig,
-};
