--- suite-native/staking/package.json
@@ -1,38 +1,0 @@
-{
-    "name": "@suite-native/staking",
-    "version": "1.0.0",
-    "private": true,
-    "license": "See LICENSE.md in repo root",
-    "sideEffects": false,
-    "main": "src/index",
-    "scripts": {
-        "depcheck": "yarn g:depcheck",
-        "type-check": "yarn g:tsc --build",
-        "test:unit": "yarn g:jest:native",
-        "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'"
-    },
-    "dependencies": {
-        "@reduxjs/toolkit": "2.12.0",
-        "@suite-common/device": "workspace:*",
-        "@suite-common/mev": "workspace:*",
-        "@suite-common/redux-utils": "workspace:*",
-        "@suite-common/wallet-config": "workspace:*",
-        "@suite-common/wallet-core": "workspace:*",
-        "@suite-common/wallet-types": "workspace:*",
-        "@suite-common/wallet-utils": "workspace:*",
-        "@suite-native/device-mutex": "workspace:*",
-        "@trezor/connect": "workspace:*",
-        "@trezor/connect-common": "workspace:*",
-        "@trezor/device-utils": "workspace:*",
-        "@trezor/env-utils": "workspace:*",
-        "@trezor/network-solana": "workspace:*",
-        "@trezor/utils": "workspace:*",
-        "react-redux": "9.3.0"
-    },
-    "devDependencies": {
-        "@suite-common/message-system": "workspace:*",
-        "@suite-common/suite-types": "workspace:*",
-        "@suite-common/test-utils": "workspace:*",
-        "@suite-native/analytics": "workspace:*"
-    }
-}
