--- suite-native/staking/src/constants.ts
@@ -1,1 +1,0 @@
-export const STAKE_NATIVE_MODULE_PREFIX = '@suite-native/staking';
