--- suite-native/staking/src/index.ts
@@ -1,15 +1,0 @@
-// Coin-specific selector files (e.g. cardanoStakingSelectors.ts) are intentionally not
-// re-exported whole. However, since selectors.ts already imports from these files,
-// they are always included in the bundle regardless. Selectors needed by external consumers
-// are therefore explicitly exported below rather than forcing deep imports.
-
-export {
-    pushStakeTransactionNativeThunk,
-    signStakeTransactionNativeThunk,
-} from './stakeNativeThunks';
-export {
-    composeSolanaStakingTransactionFeeLevelsNativeThunk,
-    signSolanaStakingTransactionNativeThunk,
-} from './stakeFormSolanaNativeThunks';
-export type * from './stakeFormSolanaNativeTypes';
-export * from './hooks/useSelector';
