--- suite-native/staking/tsconfig.json
@@ -1,44 +1,0 @@
-{
-    "extends": "../../tsconfig.base.json",
-    "compilerOptions": { "outDir": "libDev" },
-    "references": [
-        { "path": "../../suite-common/device" },
-        { "path": "../../suite-common/mev" },
-        {
-            "path": "../../suite-common/redux-utils"
-        },
-        {
-            "path": "../../suite-common/wallet-config"
-        },
-        {
-            "path": "../../suite-common/wallet-core"
-        },
-        {
-            "path": "../../suite-common/wallet-types"
-        },
-        {
-            "path": "../../suite-common/wallet-utils"
-        },
-        { "path": "../device-mutex" },
-        { "path": "../../packages/connect" },
-        {
-            "path": "../../packages/connect-common"
-        },
-        { "path": "../../packages/device-utils" },
-        { "path": "../../packages/env-utils" },
-        {
-            "path": "../../networks/solana/network-solana"
-        },
-        { "path": "../../packages/utils" },
-        {
-            "path": "../../suite-common/message-system"
-        },
-        {
-            "path": "../../suite-common/suite-types"
-        },
-        {
-            "path": "../../suite-common/test-utils"
-        },
-        { "path": "../analytics" }
-    ]
-}
